Set up a routing table, which is statically fixed at the EM service initialization

+++ services/elastic-memory/src/main/java/edu/snu/cay/services/em/evaluator/api/PartitionFunc.java
@@ -29,5 +29,5 @@ public interface PartitionFunc {
    * @param dataId a id of data
    * @return an id of partition that the data
    */
-  long partition(long dataId);
+  int partition(long dataId);
 }
