+++ services/elastic-memory/src/main/java/edu/snu/cay/services/em/evaluator/impl/RangePartitionFunc.java
@@ -39,8 +39,8 @@ public final class RangePartitionFunc implements PartitionFunc {
   }
 
   @Override
-  public long partition(final long dataId) {
-    return dataId >> partitionSizeBits;
+  public int partition(final long dataId) {
+    return (int) (dataId >> partitionSizeBits);
   }
 
   @NamedParameter(doc = "A number of bits representing partition size", default_value = "32")
